[INTERNAL,BUILD] Split PosixTarHeader-based code into seperate files
* This is to address comments on PR #1023 regarding the license of code
  taken from elsewhere that is licensed under Apache 2.
