+++ src/peakinvestigator/source/TRANSFORMATIONS/RAW2PEAK/PEAKINVESTIGATOR/PeakInvestigator.cmake
@@ -62,6 +62,10 @@ set(sources_system)
 foreach(i ${sources_system_list})
 	list(APPEND sources_system ${directory}/${i})
 endforeach(i)
+### end add path
+
+### Setup format internal source
+set(sources_format_internal ${directory}/FORMAT/INTERNAL/tarball.cpp)
 
 # header needs to be defined here, because it needs to be MOC'd
 set(peakinvestigator_header ${header_directory}/PeakInvestigator.h)
@@ -73,11 +77,13 @@ set(PeakInvestigatorFiles)
 list(APPEND PeakInvestigatorFiles ${peakinvestigator_source})
 list(APPEND PeakInvestigatorFiles ${mocced_sources})
 list(APPEND PeakInvestigatorFiles ${sources_format})
+list(APPEND PeakInvestigatorFiles ${sources_format_internal})
 list(APPEND PeakInvestigatorFiles ${sources_system})
 
 source_group("Source Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR" FILES ${peakinvestigator_source})
 source_group("Source Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR" FILES ${mocced_sources})
 source_group("Source Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\FORMAT" FILES ${sources_format})
+source_group("Source Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\FORMAT\\INTERNAL" FILES ${sources_format_internal})
 source_group("Source Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\SYSTEM" FILES ${sources_system})
 
 ## add groups for headers
@@ -105,15 +111,21 @@ set(header_system)
 foreach(i ${header_system_list})
 	list(APPEND header_system ${header_directory}/${i})
 endforeach(i)
+### end add path
+
+### Setup format internal header
+set(header_format_internal ${header_directory}/FORMAT/INTERNAL/tarball.h)
 
 list(APPEND PeakInvestigatorFiles ${peakinvestigator_header})
 list(APPEND PeakInvestigatorFiles ${header_format})
+list(APPEND PeakInvestigatorFiles ${header_format_internal})
 list(APPEND PeakInvestigatorFiles ${header_system})
 
 # define list of headers related to peakinvestigator needed for
 # installation and export
-set(PeakInvestigatorHeaders ${peakinvestigator_header} ${header_format} ${header_system})
+set(PeakInvestigatorHeaders ${peakinvestigator_header} ${header_format} ${header_format_internal} ${header_system})
 
 source_group("Header Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR" FILES ${peakinvestigator_header})
 source_group("Header Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\FORMAT" FILES ${header_format})
+source_group("Header Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\FORMAT\\INTERNAL" FILES ${header_format_internal})
 source_group("Header Files\\TRANSFORMATIONS\\RAW2PEAK\\PEAKINVESTIGATOR\\SYSTEM" FILES ${header_system})
